a whole lot
there are changes in the config files to incorportate the new menu
structure with more submenu's

there are changes in the pre_process files because some still used the
old table header function.

and in some templates there was still the hiden filed todo.
this would break the java excecution code.